발송 메일 조회(멀티건) 총 합계 단위테스트

+++ src/main/java/egovframework/com/cop/ems/service/impl/SndngMailDtlsDAO.java
@@ -40,10 +40,10 @@ public class SndngMailDtlsDAO extends EgovComAbstractDAO {
 	/**
 	 * 발송메일 총건수를 조회한다.
 	 * @param vo ComDefaultVO
-	 * @return int
+	 * @return int sndngMailsTotCnt = sndngMailDtlsDAO.selectSndngMailListTotCnt(vo);
 	 * @exception
 	 */
 	public int selectSndngMailListTotCnt(ComDefaultVO vo) {
-		return (Integer) selectOne("SndngMailDtlsDAO.selectSndngMailListTotCnt_S", vo);
+		return selectOne("SndngMailDtlsDAO.selectSndngMailListTotCnt_S", vo);
 	}
 }
